Title: 낮은 지점들, Time: 8ms, Memory: 0MB, Status: Passed - CodeTree

+++ 250104/낮은 지점들/lowest-points.cpp
@@ -18,9 +18,9 @@ int main() {
             if (y < um[x]) um[x] = y;
         }
     }
-    int ans = 0;
+    long long ans = 0;
     for (auto p : um) {
-        ans += p.second;
+        ans += (long long)p.second;
     }
     cout << ans;
     return 0;
